Fix P2P dialog time remaining text being cut off
P2P dialog time remaining text is set as a message in the AlertDialog,
separately from the ScrollView containing the rest of the P2P dialog
text. If the screen size is too small to fit the dialog and the
keyboard, the dialog will get squished and push the message underneath
the Title, cutting it off.

To prevent this, move the time remaining text to the ScrollView with the
rest of the P2P text so the text is more compact and is able to be
scrolled together if the dialog is too small.

Updated-Overlayable: Not applicable - changing default value

Bug: 288179496
Test: manually verify timeleft text appears in rest of the ScrollView
using adb shell cmd wifi launch-p2p-dialog-invitation-received
DeviceName -c 15000 -p

Change-Id: Ic2cf4dd0cbf5e33ff38023e89b3e1388cf4d1fdb